新しいAPI `PX=px2dthelper.copy_content` を追加。(未実装)
+ テストなど内部構造の整理
