+++ tests/testData/standard/copy/from.html
@@ -1,0 +1,1 @@
+<p><code>copy/from.html</code></p>
